feat: abstracted the purge logic into a separate service
