+++ src/Cogworks.CogFlare.Core/Composers/ExternalCachePurgeComposer.cs
@@ -16,6 +16,7 @@ public class ExternalCachePurgeComposer : IComposer
         builder.Services
             .AddTransient(_ => cogFlareSettings)
             .AddHttpClient()
+            .AddTransient<ICachePurgeService, CachePurgeService>()
             .AddTransient<ICloudFlareCachePurgeService, CloudFlareCachePurgeService>()
             .AddTransient<IUmbracoContentNodeService, UmbracoContentNodeService>();
